create footer and clean worksapce
