+++ src/components/Container.js
@@ -1,12 +1,6 @@
 import React from "react"
 import { ContainerWrapper } from "../elements"
-import { Header } from "./Header"
 
 export function Container({ children }) {
-  return (
-    <ContainerWrapper>
-      <Header />
-      {children}
-    </ContainerWrapper>
-  )
+  return <ContainerWrapper>{children}</ContainerWrapper>
 }
