+++ src/components/Footer.js
@@ -1,0 +1,11 @@
+import React from "react"
+import styled from "styled-components"
+
+const Footer = styled.footer`
+  padding: 50px;
+  text-align: center;
+`
+
+export function FooterContainer() {
+  return <Footer>© Hugo Meireles 2020</Footer>
+}
