--- src/elements/BlogHeaderElement.js
@@ -1,7 +1,0 @@
-import styled from "styled-components"
-
-export const HeaderContainer = styled.aside`
-  grid-column: 1 / span 12;
-
-  box-shadow: ${({ theme }) => theme.shadows.boxShadow2};
-`
