+++ src/elements/MenuElement.js
@@ -1,0 +1,11 @@
+import styled from "styled-components"
+
+export const Menu = styled.ul`
+  list-style: none;
+  display: flex;
+  align-items: center;
+
+  li {
+    padding: 10px;
+  }
+`
