+++ src/elements/index.js
@@ -1,10 +1,9 @@
 export * from "./ContainerElement"
 export * from "./SectionElement"
-export * from "./ButtonElement"
-export * from "./MenuElements"
-export * from "./NavElement"
-export * from "./LinkElement"
-export * from "./HeaderElement"
 export * from "./MainElement"
-export * from "./BlogHeaderElement"
-export * from "./ArticlesSectionElement"
+export * from "./HeaderElement"
+export * from "./ContentElement"
+export * from "./NavElement"
+export * from "./MenuElement"
+export * from "./ButtonElement"
+export * from "./LinkElement"
